Fixed potential deadlock issue
Now that we are pushing parts that may be close to the circular buffer
size, there was a potential deadlock situation that could arise:

1.  The thread pushing on the circular buffer is trying to push a buffer
that exceeds the buffer size.
2.  It queries the circular buffer to see how many it can push.
3.  Between this query more items are pushed and therefore it tries to
push beyond the circular buffer size.

To fix this, the query oh how many can be pushed needs to be in the
circular buffer itself so that it is inside the lock.  The push_back()
returns the number of bytes pushed so the caller cn determine via this
ininformation if it needs to push more.