[2146] Initialize number_of_threads in get_number_of_threads_data_size_and_opr_type
When running in the scenario described in issue 2146,
get_number_of_threads_data_size_and_opr_type is not able to determine
the number of threads as none of the ways to get it are available. Because
of this the number_of_threads remains whatever it got initialized as. In
that case the number_of_threads is likely a high number, and since this
is used to determine if shared memory can be cleaned up, the shared memory
isn't being cleaned up.  By defaulting this to zero, the shared memory gets
cleaned up properly.

Note that with this defaulting to zero we force a cache file to be used
which is fine because we don't have enough information to do otherwise.

Also note that this method is not designed to keep the initial value
passed in for number_of_threads. Defaulting to zero is what it should
have always done.